requirements.txt
langchain[llms]
python-dotenv==0.15.0
openai
langchain-openai